forecast view implemented

+++ app/src/main/java/com/github/dawidhyzy/kotlinvsjava/jv/api/model/Wind.java
@@ -8,7 +8,7 @@ public class Wind {
     @SerializedName("speed")
     private double speed;
     @SerializedName("deg")
-    private int direction;
+    private float direction;
 
     public double getSpeed() {
         return speed;
@@ -18,11 +18,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDirection() {
+    public float getDirection() {
         return direction;
     }
 
-    public void setDirection(int direction) {
+    public void setDirection(float direction) {
         this.direction = direction;
     }
